Added methods for diffing files and CLI script execution

+++ sort.py
@@ -1,5 +1,7 @@
 """Sort photos from the source directory into the destination directory."""
+import hashlib
 import shutil
+import sys
 from datetime import datetime
 from pathlib import Path
 
@@ -51,6 +53,28 @@ def is_file(file_path):
         return False
 
 
+def diff_files(src_file, dest_file):
+    """Hash two files and see if they are the same or not.
+
+    :param src_file:
+    :param dest_file:
+    :return:
+    """
+    # return run_hash(path_to_file, hashlib.sha512()).hexdigest()
+
+    src_hash, dest_hash = hashlib.sha512(), hashlib.sha512()
+
+    with open(src_file, 'rb') as file_in:
+        for block in iter(lambda: file_in.read(4096), b''):
+            src_hash.update(block)
+
+    with open(dest_file, 'rb') as file_in:
+        for block in iter(lambda: file_in.read(4096), b''):
+            dest_hash.update(block)
+
+    return src_hash.hexdigest() == dest_hash.hexdigest()
+
+
 def move_file(src_file, dest_file):
     """Move a file from the src to the dest.
 
@@ -64,7 +88,12 @@ def move_file(src_file, dest_file):
 
     dest.parent.mkdir(parents=True, exist_ok=True)
 
+    if is_file(dest) and not diff_files(src, dest):
+        return False
+
     shutil.copy2(src, dest)
+
+    return True
 
 
 def sort_images(src_path, dest_path):
@@ -92,3 +121,21 @@ def sort_images(src_path, dest_path):
         dest = dest_path / d.strftime('%Y-%m') / (prefix + d.strftime('%Y%m%d_%H%M%S') + src.suffix)
 
         move_file(src, dest)
+
+
+def main():
+    """Main method for this, typically called by
+
+    :return:
+    """
+
+    if len(sys.argv) < 3:
+        print('''Expecting two arguments:\nsource_directory destination_directory''')
+
+    src_path, dest_path = [Path(p) for p in sys.argv[1:3]]
+
+    sort_images(src_path, dest_path)
+
+
+if __name__ == '__main__':
+    main()
